change max serum3 slots to 4
